+++ app/sprinkles/core/src/Util/AssetLoader.php
@@ -8,6 +8,7 @@
  */
 namespace UserFrosting\Sprinkle\Core\Util;
 
+use UserFrosting\Assets\Assets;
 use UserFrosting\Assets\Util\MimeType;
 use UserFrosting\Support\Util\Util;
 
@@ -16,24 +17,18 @@ use UserFrosting\Support\Util\Util;
  *
  * Loads an asset from the filesystem.
  * @author Alex Weissman (https://alexanderweissman.com)
- * @deprecated
  */
 class AssetLoader
 {
-    /**
-     * @var string The base filesystem path in which to look for the asset file.  Can be an absolute path or stream.
-     */
-    protected $basePath;
-
     /**
      * @var string The fully constructed path to the file.
      */
     protected $fullPath;
 
     /**
-     * @var string A regex pattern to represent the allowed subset of paths under basePath which are accessible.
+     * @var \UserFrosting\Assets\Assets
      */
-    protected $pattern;
+    protected $assets;
 
     /**
      * Create a new AssetLoader object.
@@ -41,13 +36,9 @@ class AssetLoader
      * @param string $basePath
      * @param string $pattern
      */
-    public function __construct($basePath, $pattern)
+    public function __construct(Assets $assets)
     {
-        $this->basePath = rtrim($basePath, "/\\") . '/';
-
-        $this->pattern = $pattern;
-
-        $this->fullPath = '';
+        $this->assets = $assets;
     }
 
     /**
@@ -59,19 +50,7 @@ class AssetLoader
      */
     public function loadAsset($relativePath)
     {
-        // 1. Remove any query string
-        $relativePath = preg_replace('/\?.*/', '', $relativePath);
-
-        // 2. Normalize path, to prevent directory traversal
-        $relativePath = Util::normalizePath($relativePath);
-
-        // 3. Check that the beginning of the path matches the allowed paths pattern
-        if (!preg_match($this->pattern, $relativePath)) {
-            return false;
-        }
-
-        // 4. Build full path to file
-        $this->fullPath = $this->basePath . $relativePath;
+        $this->fullPath = $this->assets->urlPathToAbsolutePath($relativePath);
 
         // Return false if file does not exist
         if (!file_exists($this->fullPath)) {